+++ onsenui/esm/elements/ons-fab.js
@@ -63,6 +63,14 @@ export default class FabElement extends BaseElement {
    */
 
   /**
+   * @property ripple
+   * @type {Boolean}
+   * @description
+   *  [en]If this property is defined, the button will have a ripple effect when tapped.[/en]
+   *  [ja][/ja]
+   */
+
+  /**
    * @attribute position
    * @type {String}
    * @description
@@ -238,13 +246,6 @@ export default class FabElement extends BaseElement {
    *   [en]Whether the element is disabled or not.[/en]
    *   [ja]無効化されている場合に`true`。[/ja]
    */
-  set disabled(value) {
-    return util.toggleAttribute(this, 'disabled', value);
-  }
-
-  get disabled() {
-    return this.hasAttribute('disabled');
-  }
 
   /**
    * @property visible
@@ -258,6 +259,8 @@ export default class FabElement extends BaseElement {
     return this.style.transform.indexOf('scale(0)') === -1 && this.style.display !== 'none';
   }
 }
+
+util.defineBooleanProperties(FabElement, ['disabled', 'ripple']);
 
 onsElements.Fab = FabElement;
 customElements.define('ons-fab', FabElement);