+++ onsenui/esm/elements/ons-ripple.js
@@ -99,6 +99,14 @@ export default class RippleElement extends BaseElement {
    */
 
   /**
+   * @property center
+   * @type {Boolean}
+   * @description
+   *   [en]If this attribute presents, change the position of wave effect to center of the target element.[/en]
+   *   [ja]この要素を設定すると、エフェクトの位置が要素の真ん中から始まります。[/ja]
+   */
+
+  /**
    * @attribute disabled
    * @description
    *   [en]If this attribute is set, the ripple effect will be disabled.[/en]
@@ -346,14 +354,9 @@ export default class RippleElement extends BaseElement {
    *   [en]Whether the element is disabled or not.[/en]
    *   [ja]無効化されている場合に`true`。[/ja]
    */
-  set disabled(value) {
-    return util.toggleAttribute(this, 'disabled', value);
-  }
-
-  get disabled() {
-    return this.hasAttribute('disabled');
-  }
 }
+
+util.defineBooleanProperties(RippleElement, ['disabled', 'center']);
 
 onsElements.Ripple = RippleElement;
 customElements.define('ons-ripple', RippleElement);