+++ onsenui/esm/ons/util.js
@@ -39,7 +39,7 @@ platform._runOnActualPlatform(() => {
 
 try {
   const opts = Object.defineProperty({}, 'passive', {
-    get() { util.globals.supportsPassive = true; }
+    get() { util.globals.supportsPassive = true; } // eslint-disable-line getter-return
   });
   window.addEventListener('testPassive', null, opts);
   window.removeEventListener('testPassive', null, opts);
@@ -565,21 +565,25 @@ util.checkMissingImport = (...elementNames) => {
 };
 
 /**
- * On a given object, defines a boolean property that reflects an attribute of the same name.
+ * Defines a boolean property that reflects an attribute of the same name for a
+ * given list of attributes.
  */
-util.defineBooleanProperty = (object, propertyName) => {
-  Object.defineProperty(object, propertyName, {
-    get() {
-      return this.hasAttribute(propertyName);
-    },
-    set(value) {
-      if (value) {
-        this.setAttribute(propertyName, '');
-      } else {
-        this.removeAttribute(propertyName);
-      }
-    },
-    configurable: true
+util.defineBooleanProperties = (object, attributeList) => {
+  attributeList.forEach(attributeName => {
+    const propertyName = util.camelize(attributeName);
+    Object.defineProperty(object.prototype, propertyName, {
+      get() {
+        return this.hasAttribute(attributeName);
+      },
+      set(value) {
+        if (value) {
+          this.setAttribute(attributeName, '');
+        } else {
+          this.removeAttribute(attributeName);
+        }
+      },
+      configurable: true
+    });
   });
 };